+++ src/app/features/ip/types.ts
@@ -2,3 +2,4 @@ import { components, operations } from '@app/apis/schema';
 
 export type CreateNetWorkInterfaceRequestBody = components['schemas']['NetInterface'];
 export type GetNetWorkInterfaceRequestQuery = operations['netinterfaceList']['parameters']['query'];
+export type NetWorkInterface = components['schemas']['NetInterface'];