+++ src/app/features/node/components/index.ts
@@ -1,3 +1,7 @@
 export { CreateNodeForm } from './CreateNodeForm';
 
 export { VSANNodeList } from './VSANNodeList';
+
+export { Resource } from './Resource';
+
+export { StoragePool } from './StoragePool';