+++ src/app/features/report/components/index.ts
@@ -1,0 +1,2 @@
+export { CreateForm } from './CreateForm';
+export { List } from './List';