+++ src/app/interfaces/report.ts
@@ -1,6 +1,6 @@
-import { operations } from '@app/apis/schema';
+import { components, operations } from '@app/apis/schema';
 
-export type ErrorReportDeleteRequest = operations['errorReportDelete']['requestBody']['content']['application/json'];
+export type ErrorReportDeleteRequest = components['schemas']['ErrorReportDelete'];
 export type ErrorReportList = operations['errorReportList']['responses']['200']['content']['application/json'];
 
 export type ReportType = ErrorReportList[0];