+++ src/app/layouts/components/HeaderAboutModal.tsx
@@ -17,11 +17,7 @@ const HeaderAboutModal: React.FC = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [propertyModalOpen, setPropertyModalOpen] = useState(false);
 
-  const {
-    data: metrics,
-    error,
-    loading,
-  } = useRequest(fetchMetrics, {
+  const { data: metrics } = useRequest(fetchMetrics, {
     throwOnError: true,
     defaultLoading: true,
   });
@@ -67,12 +63,6 @@ const HeaderAboutModal: React.FC = () => {
     return data;
   }, [metrics]);
 
-  let VERSION = "DEV"
-
-  if (typeof process !== "undefined") {
-    VERSION = process?.env?.VERSION ?? "DEV";
-  }
-
   return (
     <React.Fragment>
       <img title="logo" className="connected__img" src={FEATHER_INFO} onClick={handleModalToggle} />
@@ -88,10 +78,10 @@ const HeaderAboutModal: React.FC = () => {
       >
         <TextContent>
           <TextList component="dl">
-            <TextListItem component="dt">Linstor Version</TextListItem>
+            <TextListItem component="dt">LINSTOR Version</TextListItem>
             <TextListItem component="dd">{linstorVersion}</TextListItem>
             <TextListItem component="dt">UI Version</TextListItem>
-            <TextListItem component="dd">{VERSION}</TextListItem>
+            <TextListItem component="dd">{process.env.VERSION}</TextListItem>
             <TextListItem component="dt">Controller Binding IP</TextListItem>
             <TextListItem component="dd">0.0.0.0</TextListItem>
             <TextListItem component="dt">Controller Active On</TextListItem>