+++ src/app/pages/Inventory/Nodes/detail.styled.ts
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  padding-bottom: 20px;
+`;
+
+export const LabelText = styled.span`
+  width: 220px;
+  font-weight: 600;
+  margin-right: 10px;
+`;
+
+export const TagContainer = styled.div`
+  line-height: 2;
+`;