+++ src/app/utils/time.ts
@@ -10,3 +10,7 @@ export const formatTime = (unixTime: number, format = 'YYYY-MM-DD HH:mm:ss'): st
   const time = dayjs(unixTime);
   return dayjs(time).format(format);
 };
+
+export const getTime = (time) => {
+  return dayjs(time).valueOf();
+};