+++ webpack.dev.js
@@ -2,14 +2,15 @@ const path = require('path');
 const { merge } = require('webpack-merge');
 const common = require('./webpack.common.js');
 const { stylePaths } = require('./stylePaths');
-const { DefinePlugin } = require('webpack');
+const dotenv = require('dotenv');
+
+dotenv.config();
 
 const HOST = process.env.HOST || '127.0.0.1';
 const PORT = process.env.PORT || '8000';
-
-const API_HOST = process.env.LISNTOR_API_HOST || 'http://192.168.123.212:3370';
-const GATEWAY_API_HOST = process.env.GATEWAY_API_HOST || 'http://192.168.123.212:8080';
-const VSAN_API_HOST = process.env.VSAN_API_HOST || 'https://192.168.123.212';
+const API_HOST = process.env.LINSTOR_API_HOST || 'http://192.168.123.214:3370';
+const GATEWAY_API_HOST = process.env.GATEWAY_API_HOST || 'http://192.168.123.214:8080';
+const VSAN_API_HOST = process.env.VSAN_API_HOST || 'https://192.168.123.214';
 
 module.exports = merge(common('development'), {
   mode: 'development',
@@ -41,11 +42,6 @@ module.exports = merge(common('development'), {
       },
     },
   },
-  plugins: [
-    new DefinePlugin({
-      'process.env.VERSION': 'DEV', // "mock" version in development environment
-    }),
-  ],
   module: {
     rules: [
       {